Add new instructions putdfloat and getdfloat to handle double floats. Update the loader to load these instructions from the xwam files. (There is a possible problem.  The doubles are written into the xwam file as bitstrings, and read by the loader as the same bitstring.  If representations on different machines are different, then xwam files compiled on one system won't run on another.)

+++ emu/emudef.h
@@ -212,17 +212,17 @@ unsigned long dec[8] = {_FULL_ULONG_BITS,_FULL_ULONG_BITS,_FULL_ULONG_BITS,_FULL
   XSB_Deref(OP1);      							\
   if (isref(OP1)) {							\
     /* op1 is FREE */							\
-      bind_boxedfloat(vptr(OP1), OP2);					\
+    bind_boxedfloat(vptr(OP1), (Float)OP2);				\
   }									\
   else if (isofloat(OP1)) {						\
-    if ( (float)ofloat_val(OP1) == OP2) {				\
+    if (ofloat_val(OP1) == (Float)OP2) {				\
       XSB_Next_Instr();							\
     }									\
     else Fail1;								\
   }									\
   else if (isattv(OP1)) {						\
     xsb_dbgmsg((LOG_ATTV,">>>> ATTV nunify_with_float, interrupt needed\n"));	\
-    bind_boxedfloat((CPtr)dec_addr(OP1), OP2);				\
+    bind_boxedfloat((CPtr)dec_addr(OP1), (Float)OP2);			\
     add_interrupt(CTXTc cell(((CPtr)dec_addr(OP1) + 1)),cell((CPtr)dec_addr(OP1))); \
   }									\
   else Fail1;	/* op1 is INT, STRING, STRUCT, or LIST */
